requirements.txt
sounddevice
scipy
numpy
openai
python-dotenv
elevenlabs
